Split user post page in two files Add test_data for data storage
